fix: update all hooks to remove use
