+++ app/hooks/events/getEventImageById.ts
@@ -1,13 +1,13 @@
 import { neon } from '@neondatabase/serverless';
 import 'server-only';
 
-export default async function useGetStartupFounderImage(startup_id: number, founder_id: number) {
+export default async function getEventImageById(event_id: number) {
   const database_url = process.env.DATABASE_URL;
 
   if (!database_url) throw new Error('DATABASE_URL null')
 
   const sql = neon(database_url);
 
-  const result = await sql`SELECT image FROM Startup WHERE startup_id = ${startup_id} AND founder_id = ${founder_id}`;
+  const result = await sql`SELECT image FROM Event WHERE id = ${event_id}`;
   return result.at(0) ?? null;
 }
