+++ app/hooks/news/getNewsById.ts
@@ -1,13 +1,13 @@
 import { neon } from '@neondatabase/serverless';
 import 'server-only';
 
-export default async function useGetPartnerById(partner_id: number) {
+export default async function getNewsById(news_id: number) {
   const database_url = process.env.DATABASE_URL;
 
   if (!database_url) throw new Error('DATABASE_URL null')
 
   const sql = neon(database_url);
 
-  const result = await sql`SELECT * FROM Partner WHERE id = ${partner_id}`;
+  const result = await sql`SELECT * FROM News WHERE id = ${news_id}`;
   return result.at(0) ?? null;
 }
