+++ app/hooks/partners/getPartnerById.ts
@@ -1,13 +1,13 @@
 import { neon } from '@neondatabase/serverless';
 import 'server-only';
 
-export default async function useGetNews() {
+export default async function getPartnerById(partner_id: number) {
   const database_url = process.env.DATABASE_URL;
 
   if (!database_url) throw new Error('DATABASE_URL null')
 
   const sql = neon(database_url);
 
-  const result = await sql`SELECT * FROM News`;
-  return result;
+  const result = await sql`SELECT * FROM Partner WHERE id = ${partner_id}`;
+  return result.at(0) ?? null;
 }
