+++ app/hooks/users/getUserByEmail.ts
@@ -1,13 +1,13 @@
 import { neon } from '@neondatabase/serverless';
 import 'server-only';
 
-export default async function useGetNewsById(news_id: number) {
+export default async function getUserByEmail(email: string) {
   const database_url = process.env.DATABASE_URL;
 
   if (!database_url) throw new Error('DATABASE_URL null')
 
   const sql = neon(database_url);
 
-  const result = await sql`SELECT * FROM News WHERE id = ${news_id}`;
+  const result = await sql`SELECT * FROM users WHERE email = ${email}`;
   return result.at(0) ?? null;
 }
