+++ app/hooks/users/getUserById.ts
@@ -1,7 +1,7 @@
 import { neon } from '@neondatabase/serverless';
 import 'server-only';
 
-export default async function useGetUserById(user_id: number) {
+export default async function getUserById(user_id: number) {
   const database_url = process.env.DATABASE_URL;
 
   if (!database_url) throw new Error('DATABASE_URL null')
