+++ app/hooks/users/getUsers.ts
@@ -1,7 +1,7 @@
 import { neon } from '@neondatabase/serverless';
 import 'server-only';
 
-export default async function useGetUsers() {
+export default async function getUsers() {
   const database_url = process.env.DATABASE_URL;
 
   if (!database_url) throw new Error('DATABASE_URL null')
